QA-405: Create a URL for directly starting a One Login journey
The stub relying party is only creating a URL based on the form inputs, so we can construct our own url as we know what the parameters will be. This bypasses a source of latency in the performance tests.

The callback url still returns to the stub as that is linked to the client id.